Avoid circular dependency warning and move the repositioning code to the component it belongs to
